Day two, step two:
Let's add another tile.  This tile will get drawn under
the mouse whenever we move it.

At this point, we're emphasizing that whenever the mouse
moves, the new tile is drawn.  But this also shows that
nothing is getting *moved*; when the mouse moves, a *new*
tile is getting painted in the new place, but there's still
a tile in the old place, too.

It's a really cool effect, and leads into another lesson -
how do we make it look like it was moved, instead of just smudged?

It's also immediately obvious that the mouse is at the upper left
of the tile, not the center.