Day two, step three:
Let's fix the alignment.

We want the tile to be centered under the mouse, not at the
corner.

The tile is 40 pixels wide and 40 pixels tall, so we need to
adjust it by half of that in each direction.

Adding 20 leads to a fun effect, which we tried just for kicks,
and then subtracting 20 got us to where we needed to be - in the
center.