Day two, cleanup:
Now that we've iterated on this a bit, let's do some cleanup.

Some of this is just changing the order and grouping of variables,
but the changes around the `one.png` and `two.png` files has a real
impact.

In the naïve implementation, we kept the tiles right next to where
they were used, and would end up having to re-read it from disk
every time.  Between the caching done by the file system and the
fact that this computer has an SSD, there wasn't any noticeable
penalty, and that's the point.  Penalties like that are often not
noticed until far too late, and of course sometimes the penalties
aren't worth spending the time to fix.

Additionally, the fix introduces python dictionaries, how to store
things in them, and how to pull them back out.